[esp/sim] move Simulator to sim namespace
Goal is to collapse Simulator and SimulatorWithAgents into one class.
Moved Simulator to sim namespace an in intermediate step. Next step
is to pull up SimulatorWithAgents logic into Simulator.

The change is part of the sensor refactoring.
