+++ src/esp/sensor/PinholeCamera.h
@@ -39,11 +39,11 @@ class PinholeCamera : public VisualSensor {
   // set the view port to the given render camera
   virtual PinholeCamera& setViewport(gfx::RenderCamera& targetCamera) override;
 
-  virtual bool getObservation(gfx::Simulator& sim, Observation& obs) override;
+  virtual bool getObservation(sim::Simulator& sim, Observation& obs) override;
 
   virtual bool getObservationSpace(ObservationSpace& space) override;
 
-  virtual bool displayObservation(gfx::Simulator& sim) override;
+  virtual bool displayObservation(sim::Simulator& sim) override;
 
   /**
    * @brief Returns the parameters needed to unproject depth for this sensor's
@@ -68,7 +68,7 @@ class PinholeCamera : public VisualSensor {
    * @param[in] sim Instance of Simulator class for which the observation needs
    *                to be drawn
    */
-  void drawObservation(gfx::Simulator& sim);
+  void drawObservation(sim::Simulator& sim);
 
   /**
    * @brief Read the observation that was rendered by the simulator
