+++ src/tests/SimTest.cpp
@@ -18,7 +18,6 @@ namespace Cr = Corrade;
 namespace Mn = Magnum;
 
 using esp::agent::AgentConfiguration;
-using esp::gfx::SimulatorConfiguration;
 using esp::nav::PathFinder;
 using esp::scene::SceneConfiguration;
 using esp::sensor::Observation;
@@ -26,6 +25,7 @@ using esp::sensor::ObservationSpace;
 using esp::sensor::ObservationSpaceType;
 using esp::sensor::SensorSpec;
 using esp::sensor::SensorType;
+using esp::sim::SimulatorConfiguration;
 using esp::sim::SimulatorWithAgents;
 
 const std::string vangogh =
